various changes

+++ portableUnitTests/OTRadValve/ThermalPhysicsModels.h
@@ -44,8 +44,8 @@ static constexpr uint_fast8_t valveUpdateTime = 60;
 // Initial conditions of the room and valve.
 struct InitConditions_t {
     // Room start temp
-    const float roomTempC;
-    const float targetTempC;
+    const double roomTempC;
+    const double targetTempC;
     // keep track of valve positions.
     const uint_fast8_t valvePCOpen;
 };
@@ -65,14 +65,14 @@ public:
      * 
      * @param   curTempC: Current temperature in C.
      */
-    virtual void tick(const float curTempC) = 0;
-    virtual void tick(const float curTempC, const uint32_t /*seconds*/) { tick(curTempC); }
+    virtual void tick(const double curTempC) = 0;
+    virtual void tick(const double curTempC, const uint32_t /*seconds*/) { tick(curTempC); }
     // Get valve percentage open.
     virtual uint_fast8_t getValvePCOpen() const = 0;
     // get target temperature in C.
-    virtual float getTargetTempC() const = 0;
+    virtual double getTargetTempC() const = 0;
     // Get the effective valve percentage open the model should use.
-    virtual float getEffectiveValvePCOpen() const = 0;    
+    virtual double getEffectiveValvePCOpen() const = 0;    
 };
 
 /**
@@ -107,7 +107,7 @@ public:
      * 
      * @param   curTempC: Current temperature in C.
      */
-    void tick(const float curTempC) override {
+    void tick(const double curTempC) override {
         is0.setReferenceTemperatures((uint_fast16_t)(curTempC * 16));
         rs0.tick(valvePCOpen, is0, NULL);
 
@@ -118,8 +118,8 @@ public:
     }
     // 
     uint_fast8_t getValvePCOpen() const override { return (valvePCOpen); }
-    float getTargetTempC() const override { return (is0.targetTempC); }
-    float getEffectiveValvePCOpen() const override { return (responseDelay.front()); }
+    double getTargetTempC() const override { return (is0.targetTempC); }
+    double getEffectiveValvePCOpen() const override { return (responseDelay.front()); }
 };
 
 
@@ -132,18 +132,20 @@ public:
 struct RoomParams_t
 {
     // Conductance of the air to the wall in W/K.
-    const float conductance_21;
+    const double conductance_21;
     // Conductance through the wall in W/K.
-    const float conductance_10;
+    const double conductance_10;
     // Conductance of the wall to the outside world in W/K.
-    const float conductance_0W;
+    const double conductance_0W;
     // Capacitance of the TODO in J/K.
-    const float capacitance_2;
+    const double capacitance_2;
     // Capacitance of the TODO in J/K.
-    const float capacitance_1;
+    const double capacitance_1;
     // Capacitance of the TODO in J/K.
-    const float capacitance_0;
+    const double capacitance_0;
 };
+
+//Modelled on DHD's office (Valve 5s, EPC Band B house).
 static const RoomParams_t roomParams_Default {
     500, 300, 50, 350000, 1300000, 7000000,
 };
@@ -154,9 +156,9 @@ static const RoomParams_t roomParams_Default {
 struct RadParams_t
 {
     // Conductance from the radiator to the room in W/K.
-    const float conductance;
+    const double conductance;
     // Maximum temperature the radiator can reach in C.
-    const float maxTemp;
+    const double maxTemp;
 };
 static const RadParams_t radParams_Default {
     25.0, 70.0
@@ -168,15 +170,15 @@ static const RadParams_t radParams_Default {
 struct ThermalModelState_t
 {
     // Inside air temperature in C
-    float airTemperature {0.0};
+    double airTemperature {0.0};
     // ??
-    float roomTemp {0.0};
-    float t1 {0.0};
-    float t0 {0.0};
+    double roomTemp {0.0};
+    double t1 {0.0};
+    double t0 {0.0};
     // Temperature of the outside world in C.
-    float outsideTemp {0.0};
+    double outsideTemp {0.0};
     // Temperature at the rad valve in C.
-    float valveTemp {0.0};
+    double valveTemp {0.0};
 };
 
 static void initThermalModelState(ThermalModelState_t& state, const InitConditions_t& init) {
@@ -196,8 +198,6 @@ static void initThermalModelState(ThermalModelState_t& state, const InitConditio
  * 
  * Additionally, heat flow to the radvalve is modelled to allow simulating its
  * position.
- * 
- * Modelled on DHD's office (Valve 5s, EPC Band B house).
  */
 class ThermalModelBasic
     {
@@ -210,13 +210,13 @@ class ThermalModelBasic
         const RoomParams_t roomParams;
         const RadParams_t radParams;
 
-        float radHeatFlow {0.0f};
+        double radHeatFlow {0.0};
 
         // Internal methods
         /**
          * @brief   Calculate heat transfer through a thermal resistance. Flow from temp1 to temp2 is positive.
          */
-        static float heatTransfer(const float conductance, const float temp1, const float temp2)
+        static double heatTransfer(const double conductance, const double temp1, const double temp2)
         {
             return conductance * (temp1 - temp2);
         }
@@ -229,12 +229,12 @@ class ThermalModelBasic
          *          - Cannot be below air temperature (the radiator cannot sink heat).
          * @retval  Heat transfer into room from radiator, in J
          */
-        float calcHeatFlowRad(const float airTemp, const uint8_t radValveOpenPC) const 
+        double calcHeatFlowRad(const double airTemp, const uint8_t radValveOpenPC) const 
         {
             // convert radValveOpenPC to radiator temp (badly)
-            const float radTemp = (2.0 * (float)radValveOpenPC) - 80.0;
+            const double radTemp = (2.0 * (double)radValveOpenPC) - 80.0;
             // Making sure the radiator temp does not exceed sensible values
-            const float scaledRadTemp = (radTemp < radParams.maxTemp) ? radTemp : radParams.maxTemp;
+            const double scaledRadTemp = (radTemp < radParams.maxTemp) ? radTemp : radParams.maxTemp;
             // Calculate heat transfer, making sure rad temp cannot go below air temperature.
             return (radTemp > airTemp) ? (heatTransfer(radParams.conductance, scaledRadTemp, airTemp)) : 0.0;
         }
@@ -247,14 +247,14 @@ class ThermalModelBasic
          *          - Cannot be below air temperature (the radiator cannot sink heat).
          * @retval  Heat transfer into room from radiator, in J
          */
-        float calcValveTemp(const float airTemp, const float localTemp, const float heatFlowFromRad) const
+        double calcValveTemp(const double airTemp, const double localTemp, const double heatFlowFromRad) const
         {
-            static constexpr float thermalConductanceRad {0.05f};  // fixme literal is starting estimate for thermal resistance
-            static constexpr float thermalConductanceRoom {10.0f};
-            const float heatIn = heatFlowFromRad * thermalConductanceRad;
-            const float heatOut = heatTransfer(thermalConductanceRoom, localTemp, airTemp);
-            const float valveHeatFlow = heatIn - heatOut;
-            const float newLocalTemp = localTemp + (valveHeatFlow / 5000);  // fixme literal is starting estimate for thermal capacitance
+            static constexpr double thermalConductanceRad {0.05};  // fixme literal is starting estimate for thermal resistance
+            static constexpr double thermalConductanceRoom {10.0};
+            const double heatIn = heatFlowFromRad * thermalConductanceRad;
+            const double heatOut = heatTransfer(thermalConductanceRoom, localTemp, airTemp);
+            const double valveHeatFlow = heatIn - heatOut;
+            const double newLocalTemp = localTemp + (valveHeatFlow / 5000);  // fixme literal is starting estimate for thermal capacitance
             return newLocalTemp;
         }
 
@@ -280,23 +280,23 @@ class ThermalModelBasic
         void calcNewAirTemperature(uint8_t radValveOpenPC) {
             radValveInternal.set(radValveOpenPC);
             // Calc heat in from rad
-            const float heat_in = calcHeatFlowRad(roomState.airTemperature, radValveInternal.get());
+            const double heat_in = calcHeatFlowRad(roomState.airTemperature, radValveInternal.get());
             radHeatFlow = heat_in;
 
             // Calculate change in heat of each segment.
-            const float heatDelta_21 = heatTransfer(roomParams.conductance_21, roomState.roomTemp, roomState.t1);
-            const float heatDelta_10 = heatTransfer(roomParams.conductance_10, roomState.t1, roomState.t0);
-            const float heatDelta_0w = heatTransfer(roomParams.conductance_0W, roomState.t0, roomState.outsideTemp);
+            const double heatDelta_21 = heatTransfer(roomParams.conductance_21, roomState.roomTemp, roomState.t1);
+            const double heatDelta_10 = heatTransfer(roomParams.conductance_10, roomState.t1, roomState.t0);
+            const double heatDelta_0w = heatTransfer(roomParams.conductance_0W, roomState.t0, roomState.outsideTemp);
 
             // Calc new heat of each segment.
-            const float heat_21 = heat_in - heatDelta_21;
-            const float heat_10 = heatDelta_21 - heatDelta_10;
-            const float heat_out = heatDelta_10 - heatDelta_0w;
+            const double heat_21 = heat_in - heatDelta_21;
+            const double heat_10 = heatDelta_21 - heatDelta_10;
+            const double heat_out = heatDelta_10 - heatDelta_0w;
 
             // Calc new temps.
-         roomState.roomTemp += heat_21 / roomParams.capacitance_2;
-         roomState.t1 += heat_10 / roomParams.capacitance_1;
-         roomState.t0 += heat_out / roomParams.capacitance_0;
+            roomState.roomTemp += heat_21 / roomParams.capacitance_2;
+            roomState.t1 += heat_10 / roomParams.capacitance_1;
+            roomState.t0 += heat_out / roomParams.capacitance_0;
 
             // Calc temp of thermostat. This is the same as the room temp in a splot unit.
             if(!splitUnit) { roomState.valveTemp = calcValveTemp(roomState.roomTemp, roomState.valveTemp, heat_in); }
@@ -304,7 +304,7 @@ class ThermalModelBasic
             if(verbose) { }  // todo put print out in here
         }
         ThermalModelState_t getState() const { return  (roomState); }
-        float getHeatInput() const { return (radHeatFlow); }
+        double getHeatInput() const { return (radHeatFlow); }
     };
 
 /**
@@ -315,7 +315,7 @@ class ThermalModelBasic
  * @param   targetTempC: target room temperature (key 'tT|C').
  * @param   valvePCOpen: current valve position in % (key 'v|%').
  */
-static void printFrame(const unsigned int i, const ThermalModelState_t& state, const float targetTempC, const uint_fast8_t valvePCOpen) {
+static void printFrame(const unsigned int i, const ThermalModelState_t& state, const double targetTempC, const uint_fast8_t valvePCOpen) {
     // fprintf(stderr, "[ \"%u\", \"\", {\"T|C\": %.2f, \"TV|C\": %.2f, \"tT|C\": %.2f, \"v|%%\": %u} ]\n",
     //         i, state.airTemperature, state.valveTemp, state.targetTemp, valvePCOpen);
     fprintf(stderr, "[ \"%u\", \"\", {\"T|C\": %.2f, \"TV|C\": %.2f, \"tT|C\": %.2f, \"v|%%\": %u} ]\n",
@@ -327,9 +327,9 @@ struct TempBoundsC_t {
     // Delay in minutes to wait before starting to record values.
     const uint32_t startDelayM = 100;
     // Maximum temperature observed in C
-    float max = 0.0;
+    double max = 0.0;
     // Minumum temperature observed in C
-    float min = 100.0;
+    double min = 100.0;
 };
 
 /**
@@ -338,7 +338,7 @@ struct TempBoundsC_t {
  * @param   bounds: Previous min and max temps.
  * @param   roomTemp: Current room temp
  */
-static void updateTempBounds(TempBoundsC_t& bounds, const float roomTemp)
+static void updateTempBounds(TempBoundsC_t& bounds, const double roomTemp)
 {
     bounds.max = (bounds.max > roomTemp) ? bounds.max : roomTemp;
     bounds.min = (bounds.min < roomTemp) ? bounds.min : roomTemp;
